requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
MySQL-python==1.2.5
lxml==3.4.2
